+++ packages/crypto-proposal/test/fixtures/index.ts
@@ -1,3 +1,3 @@
 export * from "./block.js";
 export * from "./lock-proof.js";
-export * from "./proposal.js";
+export * as Proposal from "./proposal.js";
