+++ packages/crypto-proposal/test/fixtures/proposal.old.ts
@@ -1,0 +1,80 @@
+import type { Contracts } from "@mainsail/contracts";
+
+import { blockHeader, serializedBlock } from "./block.js";
+import { serializedLockProof } from "./lock-proof.js";
+
+// Scenarios:
+// 1. Proposal without valid round
+// 2. Proposal with valid round
+// 3. Proposal without valid round, with lock proof
+// 4. Proposal with valid round, with lock proof
+
+// Names:
+// 1. serializedProposal
+// 2. serializedProposalWithValidRound
+// 3. serializedProposalWithLockProof
+// 4. serializedProposalWithValidRoundWithLockProof
+
+
+// PROPOSAL WITHOUT VALID ROUND
+export const proposalDataUnsigned: Contracts.Crypto.SerializableProposalData = {
+	data: {
+		serialized: serializedBlock,
+	},
+	round: 1,
+	signature:
+		"b7010f03f72afb5437da8f7ee039a7fee75d6e9c7b02e1b9cbd4ce844cdc0e81233fd312cdd493e4ef2c2a6ac3c9fc8a1967f06a1a205c3daf369ac77f0a895717c520af5e341a3925d23b126d847a6fd1e194a010b89082039e1e5b44352616",
+	validRound: undefined,
+	validatorIndex: 0,
+};
+
+export const proposalData: Contracts.Crypto.ProposalData = {
+	blockHeader,
+	data: {
+		serialized: serializedBlock,
+	},
+	lockProof: undefined,
+	round: 1,
+	signature:
+		"b7010f03f72afb5437da8f7ee039a7fee75d6e9c7b02e1b9cbd4ce844cdc0e81233fd312cdd493e4ef2c2a6ac3c9fc8a1967f06a1a205c3daf369ac77f0a895717c520af5e341a3925d23b126d847a6fd1e194a010b89082039e1e5b44352616",
+	validRound: undefined,
+	validatorIndex: 0,
+};
+
+export const serializedProposal =
+	"0100000000b0020000" + serializedBlock + "00b7010f03f72afb5437da8f7ee039a7fee75d6e9c7b02e1b9cbd4ce844cdc0e81233fd312cdd493e4ef2c2a6ac3c9fc8a1967f06a1a205c3daf369ac77f0a895717c520af5e341a3925d23b126d847a6fd1e194a010b89082039e1e5b44352616";
+
+export const serializedProposalUnsigned =
+	"0100000000b0020000" + serializedBlock + "00";
+
+// PROPOSAL WITH VALID ROUND
+export const proposalDataWithValidRound: Contracts.Crypto.ProposalData = {
+	blockHeader,
+	data: {
+		serialized: serializedBlock,
+	},
+	lockProof: undefined,
+	round: 1,
+	signature:
+		"00b25fd16693a2246d3e9dff3d7ae3da1473c1b24f6ef8b33c69e38e21e32c9ca307ae5f7d4573e2679ec48ecc2f1ff89d16115d8e7f6ffcba72cc9a746bcc40cdd0f9120d87c20addb55baceea2cb8cdb75faaf036e6a4221d28dc8f6558d8c",
+	validRound: 0,
+	validatorIndex: 0,
+};
+
+export const serializedProposalWithValidRound =
+	"010000000100000000b0020000000174ba618a8c01020000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200d007000000000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000000000000000000000e2000000f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0fb559f4fbb75c378cdd3dd7ccbfeff9c5c2094e5569000000f8678085012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a0921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768a022db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd69000000f8670185012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a06c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fca01a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d00000000000000000000b25fd16693a2246d3e9dff3d7ae3da1473c1b24f6ef8b33c69e38e21e32c9ca307ae5f7d4573e2679ec48ecc2f1ff89d16115d8e7f6ffcba72cc9a746bcc40cdd0f9120d87c20addb55baceea2cb8cdb75faaf036e6a4221d28dc8f6558d8c";
+
+export const serializedProposalDataWithValidRoundUnsigned =
+	"010000000100000000b0020000000174ba618a8c01020000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200d007000000000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000000000000000000000e2000000f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0fb559f4fbb75c378cdd3dd7ccbfeff9c5c2094e5569000000f8678085012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a0921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768a022db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd69000000f8670185012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a06c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fca01a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d000000000000000000";
+
+
+// TODO: Fix data
+export const serializedBlockWithLockProof =
+	"000174ba618a8c01020000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200d007000000000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000000000000000000000e2000000f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0fb559f4fbb75c378cdd3dd7ccbfeff9c5c2094e5569000000f8678085012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a0921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768a022db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd69000000f8670185012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a06c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fca01a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d0000000000000000";
+
+export const serializedProposalData = "00" + serializedBlock;
+
+export const serializedProposedDataWithLockProof = "69" + serializedLockProof + serializedBlock;
+
+
+
